refactor: MVC 패턴 적용
