+++ src/main/java/lotto/controller/LottoGameController.java
@@ -1,0 +1,4 @@
+package lotto.controller;
+
+public class LottoGameController {
+}
