+++ src/main/java/lotto/view/View.java
@@ -1,0 +1,4 @@
+package lotto.view;
+
+public class View {
+}
